requirement.txt
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
blinker==1.6.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
crontab==1.0.1
cryptography==41.0.7
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2
django-crontab==0.7.1
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
Flask==3.0.0
future==0.18.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
Markdown==3.5
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
Serializer==0.2.1
serializers==0.2.4
simplejson==3.19.2
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
views==0.3
Werkzeug==3.0.0
